uploaded images for design page moved images for all pages to pageNamed folders in images, adjusted import paths
