+++ src/components/icons.js
@@ -1,8 +1,8 @@
-import DribbbleIcon from    '../../public/images/dribbble-icon.png';
-import LetterIcon from      '../../public/images/letter-icon.png';
-import LinkedInIcon from    '../../public/images/linkedin-icon.png';
-import DownloadIcon from    '../../public/images/download-icon.png';
-import CopyIcon from      '../../public/images/copy-icon.png';
+import DribbbleIcon from    '../../public/images/icons/dribbble-icon.png';
+import LetterIcon from      '../../public/images/icons/letter-icon.png';
+import LinkedInIcon from    '../../public/images/icons/linkedin-icon.png';
+import DownloadIcon from    '../../public/images/icons/download-icon.png';
+import CopyIcon from      '../../public/images/icons/copy-icon.png';
 
 
 export {
