+++ src/components/resumeImages.js
@@ -1,5 +1,5 @@
-import ResumePage1 from '../../public/images/resume_pg1.jpg'
-import ResumePage2 from '../../public/images/resume_pg2.jpg'
+import ResumePage1 from '../../public/images/resume/resume_pg1.jpg'
+import ResumePage2 from '../../public/images/resume/resume_pg2.jpg'
 
 export {
     ResumePage1,
